added cli image and stopchecks, changed path env in dockerfile instead of bashrc to have binaries in path on other shells as well

+++ base/provisioners/01_composer_install.sh
@@ -1,2 +1,11 @@
 #!/bin/bash
-php $PROJECT_ROOT/composer.phar install --optimize-autoloader --no-progress --no-interaction --no-ansi --prefer-dist
+echo "Composer install provisioner start. Searching for composer in $PROJECT_ROOT..."
+if [ -e $PROJECT_ROOT/composer.phar ]; then
+  php $PROJECT_ROOT/composer.phar install --optimize-autoloader --no-progress --no-interaction --no-ansi --prefer-dist
+fi;
+echo "Composer executable not found in $PROJECT_ROOT, searching if composer is installed globally..."
+EXECUTABLE=$(which composer)
+if [ -x "$EXECUTABLE" ]; then
+   $EXECUTABLE install --optimize-autoloader --no-progress --no-interaction --no-ansi --prefer-dist
+fi
+echo "No Composer executable found, skip dependency installation."
